dom/alarm/test/test_alarm_add_date.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test date Paramter for Alarm API</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
  <script type="application/javascript">

  "use strict";

  // Verify passing a Date in the future doesn't fail
  function testFutureDate() {
    var tomorrow = new Date();
    tomorrow.setDate(tomorrow.getDate() + 1);

    var domRequest;
    try {
      domRequest = navigator.mozAlarms.add(tomorrow, "honorTimezone", {});
    } catch (e) {
      ok(false,
         "Unexpected exception trying to add alarm for tomorrow.");

      // Proceed to next test.
      return testPastDate();
    }
    domRequest.onsuccess = function(e) {
      navigator.mozAlarms.remove(e.target.result);
      ok(true, "Add alarm for future date.");

      // Awesome, no error so proceed to next test.
      testPastDate();
    };
    domRequest.onerror = function(e) {
      ok(false, "Unable to add alarm for tomorrow`.");

      // Proceed to next test.
      testPastDate();
    };
  }

  // Verify passing a Date that's already past doesn't fail (it should fire immediately).
  function testPastDate() {
    var yesterday = new Date();
    yesterday.setDate(yesterday.getDate() - 1);

    var domRequest;
    try {
      domRequest = navigator.mozAlarms.add(yesterday, "honorTimezone", {});
    } catch (e) {
      ok(false,
         "Unexpected exception trying to add alarm for yesterday.");

      // Move on to the next test.
      return testNullDate();
    }
    domRequest.onsuccess = function(e) {
      navigator.mozAlarms.remove(e.target.result);

      ok(true, "Should be able to add alarm for already past date, which should fire immediately.");

      // Move on to the next test.
      testNullDate();
    };
    domRequest.onerror = function(e) {
      ok(false, "Unable to add alarm for yesterday.");

      // Move on to the next test.
      testNullDate();
    }
  }

  // Verify passing null does indeed fail
  function testNullDate() {
    try {
      navigator.mozAlarms.add(null, "honorTimezone", {});
      ok(false, "Expected an exception to be thrown for alarm with null date.");
    } catch(e) {
      ok(true, "Exception thrown for alarm with null date.");
    }

    // Move on to the next test.
    testInvalidTimeZone()
  }

  function testInvalidTimeZone() {
    try {
      navigator.mozAlarms.add(new Date(), "badTimeZoneArg", {});
      ok(false, "Expected an exception to be thrown while testing bad time zone arg.");
    } catch(e) {
      ok(true, "Exception thrown while testing bad time zone arg.");
    }
    SimpleTest.finish();
  }

  function startTests() {
    SpecialPowers.pushPrefEnv({
      "set": [["dom.mozAlarms.enabled", true]]
    }, function() {
      var isAllowedToTest = true;

      if (navigator.appVersion.indexOf("Android") !== -1) {
        ok(true, "mozAlarms is not allowed on Android for now. " +
                 "TODO Bug 863557.");
        isAllowedToTest = false;
      } else if (SpecialPowers.wrap(document).nodePrincipal.appStatus ==
                 SpecialPowers.Ci.nsIPrincipal.APP_STATUS_NOT_INSTALLED) {
        ok(true, "mozAlarms is not allowed for non-installed apps. " +
                 "TODO Bug 876981.");
        isAllowedToTest = false;
      }

      if (isAllowedToTest) {
        ok(true, "Start to test...");
        testFutureDate();
      } else {
        // A sanity check to make sure we must run tests on Firefox OS (B2G).
        if (navigator.userAgent.indexOf("Mobile") != -1 &&
            navigator.appVersion.indexOf("Android") == -1) {
          ok(false, "Should run the test on Firefox OS (B2G)!");
        }

        SimpleTest.finish();
      }
    });
  }

  SimpleTest.expectAssertions(0, 9);
  SimpleTest.waitForExplicitFinish();
  if (SpecialPowers.hasPermission("alarms", document)) {
    startTests();
  } else {
    // Add the permissions and reload so they propogate
    SpecialPowers.addPermission("alarms", true, document);
    window.location.reload();
  }

  </script>
</pre>
</body>
</html>
